fix: resolve C4267 warning in MftReader.cpp
- Change `globalToLocal` key type from `uint32_t` to `size_t` to prevent data loss warnings on x64.
- Ensure type consistency in drive cache serialization logic.
- Achieve zero-warning build state for ArcMeta project.

+++ src/mft/MftReader.cpp
@@ -325,7 +325,7 @@ bool MftReader::saveDriveToCacheInternal(size_t driveIdx) {
     std::vector<uint32_t> no, attr, ds;
     std::vector<uint8_t> sp, mf;
     std::unordered_map<uint32_t, uint32_t> offsetMap;
-    std::unordered_map<uint32_t, uint32_t> globalToLocal;
+    std::unordered_map<size_t, uint32_t> globalToLocal; // 2026-05-14 修正：使用 size_t 消除 C4267 警告
 
     for (size_t i = 0; i < m_frns.size(); ++i) {
         if (m_frns[i] != 0 && (m_parent_frns[i] >> 48) == driveIdx) {
